Group savings plan/Api response/Use of Dto in request

See merge request tm30/adashi/backend!13

+++ app/Domain/Dto/Request/Integrations/Paystack/PaystackPaymentRequestDto.php
@@ -12,7 +12,7 @@ class PaystackPaymentRequestDto extends PaymentProviderRequestDto
     public function __construct(
         string $authorizationCode,
         string $reference,
-        int $amount,
+        float $amount,
         string $email
     ) {
         $this->authorizationCode = $authorizationCode;
